Cleanup noise diode timestamp logs
Default log output per antenna when setting the noise diode is cluttering the output logs
The new structure will note if the noise diode set times returned by the antennas are different and
alert the user.
Else if all timestamps are in sync, the script will use the average of the timestamp in the log
message reported to the user.