Improve slew time estimation for simulated observations (#72)
* saving initial tests on astrokat simulated slew times

* printing the az-el distance to slew instead of the coordinates from point A to point B

* adding the estimated overhead time of 2.5 seconds

* Adjsuting simulation slew time to be that of an ideal telescope system

* logging the coordinates to the next target

* Adjusting total observation times to match the adjustments made in the slewing times

* Moving 3 second lag time

The 3 second delay introduced by setting up the telescope must be in 
`simulate.py`

* Addressing PR Comments: Updated Jenkinsfile accomodating tox changes and updated simulate module

* Updated to accomodate tox changes

* Updated tox.ini to check test coverage and added dependencies from setup.py

* setup.py: updated classifiers, tox.ini: added the [tox] which was accidentally removed

* Renaming

Renamed method that was previously `slew_time` to `slew_time_details` 
and removed the handling of a possible `AttributeError` As these are not 
likely and `mkat_session` implementation does not consider them.

* Removing unused implementation

* Fixing typos

Formatting comments and fixing spelling

* Consistent call for slew_time_details method

We made `_fake_slew()` return as `slew_time_details()`,  and are calling it once when we use it later.

* Rework calculation of simulated target az and el

Grouped method related to fake slew and calculations together, and made them "private" methods,
as they do not need to be exposed to users of session.

Fixed logging of target az and el.  It is now available first the first target as well.

JIRA: MT-1034

* Simplify updating of simulated time

Rather than updating `simobserver` after each call to `time.sleep`, move the update into the
simulated sleep method.

JIRA: MT-1034

* Defined capture init method

Capture init method to log the time we start simulated observation.
This method does the 3 seconds wait dedicated to observation setup in real observations.
We want to simulate the wait.