Add unit tests for simulated slew
This is in preparation for improving the slew speed calculations.

Moved `katpt_current` target update from `track` to `_fake_slew` as
it is not specific to track.  Other types of motion like raster scan
could use fake slew as well, in future.  It also makes testing easier.

Removed unittest2 as it is only useful for Python < 2.7.