OPS-1899 Increase noise diode lead time to 5 s
Turning off 63 L-band noise diodes on 2021-04-15 took about 3.3 seconds,
much longer than the usual 1.0 - 1.5 seconds. This caused the final few
KATCP requests to fail and the script to crash.

The culprit was probably the increased number of S-band digitisers
being serviced by the digitiser master controller, so this situation
may persist in the near future.

For now, increase the lead time from 3.0 to 5.0 seconds. Fix the unit
test to use a specific lead time to avoid dependence on the default
value.