OPS-1899 Move lead time constant down to module
The _DEFAULT_LEAD_TIME constant is seemingly only exposed at the top
level so that it can be passed to observe_main, but it makes more sense
to have its natural home in the `noisediode` module. It seems awkward
for `noisediode` to import its own constant from the public API,
instead of exporting it the other way.

Instead, indicate the default lead time by the value `None`, and let
the lower-level functions substitute it for _DEFAULT_LEAD_TIME. This
also avoids exposing an underscored "private" attribute in the public
API.

By the same token, move `max_cycle_len` to `noisediode` as well. This
case is even more straightforward, since there are no apparent use cases
outside of `noisediode`.