OPS-1899 Handle too short ND lead times better
Currently, if the lead time is too short, the noise diode request will
fail. This results in an empty `timestamps` list in `_set_dig_nd_`,
which then turns into a NaN via `np.mean` and this finally crashes the
script when turned into an int in the debug log message. This means
that a single noise diode request failure will crash the script.

Since we know that a requested timestamp will be out of date beforehand,
check for that, report the error and skip the rest of the noise diodes.
Assemble the replies instead of the timestamps, and process them in one
go (since `_katcp_reply_` already handles a sequence). Suppress the
"mean of empty slice" error message by short-circuiting that case
directly to NaN. Only report the timestamp if it is not NaN.