OPS-1899 Improve logging output
Use katpoint.Timestamp to have an easy way to display milliseconds from
a Unix timestamp. A display resolution of seconds may otherwise obscure
why the lead time was too small. Also print the timestamp in the same
common style as script log times instead of the `time.ctime` style.

Join the skipped antenna names with ',' instead of ', ' to be consistent
with other displays of antenna name lists. Don't use the word "on" in
the error message, since it is confusing.