OPS-1899 Better log message for failed ND request
Currently a failed noise diode request logs the following:

    WARNING  Unexpected noise diode reply from ant m005

This not very helpful. While more info is available in debug logs,
the log level is typically not DEBUG during an observation and
switching it is a hassle (and the problem might not come back on a
second observation). The reply is also not "unexpected"; the request
simply failed.

Put all pertinent information in the warning message. Now that we skip
outdated noise diode requests, we should not be swamped with loads of
failed requests.

Two style improvements: use the `reply.succeeded` property as used by
CaptureSession, and don't `continue` a for-loop if that is the final
action of the loop.