Support pyephem 4.0 time parser
Remove timezone info from `start_time` after loading
from YAML.  If times have a trailing "Z", then
the string representations are like 2019-02-11 02:10:47+00:00.  ephem < 3.7.7.1 was OK to
parse that (and ignored the timezone suffix).  Since
version 4.0.0.1, the ephem parser is stricter and
complains about these times.

This problem wasn't seen in the astrokat unit tests
for this repo, but was picked up in the CAM AQF
tests.  CAM AQF use katcorelib, instead of the simulated
session.  The difference in katcorelib is that the
start time is converted via `katpoint.Timestamp`,
instead of `astrokat.datetime2timestamp`.

Added a unit test to ensure the conversions between
astrokat and katpoint are equivalent, and exercise
the ephem converter.