Fix Timestamp issue on driftscan
Adding an float  to a ephem_date date object returns an float number of days from 1900 plus the float. This when passed to a katpoint target object is interperated as a unix epoch,  number of seconds from 1970.  Two bugs in the code : First the duration that is added is in seconds but when added to ephem_date is interperated as days. Second the result which is days is assumed to be unix epoch seconds. This is fixed by first converting the ephem_date object to a katpoint.Timestamp object then adding the float number of seconds. Then using  the to_ephem_date method of the katpoint.Timestamp to convert it back.