SPP1-168: Allow different targets to be set
Created the  ra dec target to be set on CBF for delay tracking.
The interface to `CaptureSession.set_target` will accept a dictionary of targets instead of a single target,
in `reversescan` we create a simple target dictionary which contain the following keys: `cbf`, `ants` with
respectic `scan_nd_az_el` for the area extent scan and `radec target`
for delay tracking on CBF.

To maintain the code flexibility we use an `instance` checks for the
target dictionary to access the `scan_target` or `radec target`  when required by the reverse scan simulations, while not affecting the other scans in `astrokat`.